requirements.txt
streamlit==1.45.1
jira==3.8.0
openai==1.82.1
python-dotenv==1.1.0
langchain==0.3.25
langchain-openai==0.3.18
langchain-community==0.3.18